// frontend/src/api-client/models/RecipeCreationDto.ts
/* generated using openapi-typescript-codegen -- do not edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */
import type { RecipeIngredientDTO } from './RecipeIngredientDTO';
export type RecipeCreationDto = {
    userId?: number;
    name?: string | null;
    ingredients?: Array<RecipeIngredientDTO> | null;
};